make santized type a classmethod
